+++ base/cxx/nsocket/nsocket/http/server.cpp
@@ -4,7 +4,7 @@ namespace nsocket
 {
 namespace http
 {
-Server::Server(const std::string& name, size_t threadCount, const std::string& host, unsigned int port, bool reuseAddr, size_t bz,
+Server::Server(const std::string& name, size_t threadCount, const std::string& host, uint16_t port, bool reuseAddr, size_t bz,
                size_t handshakeTimeout)
 {
     m_tcpServer = std::make_shared<TcpServer>(name, threadCount, host, port, reuseAddr, bz, handshakeTimeout);
